pyproject.toml
[project]
name = "weather-checker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "streamlit (>=1.47.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pytz (>=2025.2,<2026.0)"
]

[tool.poetry]
packages = [{include = "weather_checker", from = "src"}]


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
